// admin-clients/libs/cpanel/sales/data-access/src/lib/sales/orders/models/order-filter-fields.enum.ts
export enum OrderFilterFields {
    channels = 'channels',
    events = 'events',
    sessions = 'sessions',
    eventEntity = 'event-entities',
    channelEntity = 'channel-entities',
    merchants = 'merchants',
    users = 'users',
    clients = 'clients',
    deliveryMethod = 'delivery-method',
    currencies = 'currencies'
}